Change mdbootstrap version
A newer version of mdbootstrap has been released
post Beijing release. This new version has an error
which is causing a jenkins build failure.

The version is being set to the version that was
used at the time of the Beijing release (4.5.3)
in order to fix the failing build.

The code will still match the released artifacts,
since version 4.5.3 is contained in the release.

Change-Id: I45545119ff127c20d588be5fd08c6b367d18de03
Signed-off-by: Patrick Brady <pb071s@att.com>
Issue-ID: APPC-1072
(cherry picked from commit d081361033fc48f12ce95d3b36717ceca0abf10c)