Update patch set 2
Patch Set 2:

Oleg, these CSS classes that i have added in styles.css, they are specific to the ace editor used in template and param modules. They belong to the ace editor provided by npm. I am just overriding the color of text selection. So, now when we select any text inside the ace editor in template or param module,it will be selected in blue color. Earlier, it was grey and because of that, it was not visible clearly whether the text has been selected or not.

Patch-set: 2