Update patch set 3
Patch Set 3:

Oleg, the introduction of this class is quite beneficial in the long run, when more APIs come into picture. This new file will serve as a common file to manage the API calls. For example, while doing a save all now in the reference page, we are calling the getArtifact API 4 times, one after the other. So, in a way we are writing the same piece of code4 times, which results in redundancy and in the future, it doesnt help in scalability as well. 
Once, this file is merged, i will make changes to the application and call all the APIs through this file itself. I already have the code ready and regression tested.

Patch-set: 3