Update patch set 3
Patch Set 3:

In this we are testing the 'collapse' get method, that returns the 'isExpanded' value. 
We are setting the isExpanded variable to a boolean value and then expecting the 'collapse' get method to return the same value. 
This will test the 'collapse' method for correct return value and also cover the method in code coverage.

Patch-set: 3
Reviewer: Gerrit User 4251 <4251@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0