Update patch set 4
Patch Set 4: -Code-Review

I tried again to add this new style in the root style sheet file (src/styles.css), rebuild the CDT GUI, start it in Firefox, upload a template from my PC to the browser, see it in the ace editor, tried to select some text, but the background of the selected text is still grey. Don't know how to test the color.

Patch-set: 4