parameter.component - fixed test case
fixed test case 'should retrive PD from cache' that was failing with the
below error as it was expecting wrong value.
Error: Expected [ ] to be 'TEST PD INFORMATION'

In this test case, ngAfterViewInit() method is called which returns the
PD data. This PD data is stored into variable called
'displayParamObjects' and then this variable is returned to the calling
function.

The value for 'displayParamObjects' is set in getPDFromSession() method
which makes a async http call to get the PD data, so even before the
results of the async call are returned, ngAfterViewInit() method
returns the empty 'displayParamObject' variable.

Hence changed the expected value to the actual value returned by
ngAfterViewInit() method.

We can write a seperate test case to test the async call made by
getPDFromSession() method.

Issue-ID: APPC-1064
Change-Id: I0b8f38b6e52dc7478a4ab510ca3ff10ff81ae0ca
Signed-off-by: Arundathi Patil <arundpil@in.ibm.com>